requirements.txt

giza_actions==0.3.0
giza_cli==0.15.0
giza_agents
giza-sdk = "^0.1.1"
numpy==1.26.4
pandas==1.5.3
prefect==2.14.6
python-dotenv==1.0.1
scikit_learn==1.4.2
torch==2.2.2